Base commit of the Altia Connect (casting) R&D application. Things will be changing fairly frequently, as this is a prototype at the moment
To get started, there shouldn't be much that needs to be done for both Windows and Linux. Qt is being used for the development, but that will switch to an altia driven interface once a little more work is done on routing the media output. Qt Creator 3.6 (based on Qt 5.5.1) is being used and unless there are any hardcoded paths that I did not catch, the project should be able to compile out of the box.

As a note, the Debug_x86_64 debug/release folder contains two very important elements that need to be transferred to any build folder (where the exe exists). The first is the config.ini file. This one contains the settings for the webserver that is part of this application. The second is the docroot folder that contains all of the xml/device/scpd description files along with icons etc..

As things get a little more stable, I will clean up resource locations and consolidate things down a bit more (there is a bunch of 'fluff' in this app atm)

Once everything is compiled, the server will start on its own (view 'Settings' tab), and the user can then use the AllCast application to broadcast Pictures, Music, and Media to the screen.
Where this project enters more of the "R"esearch aspect, is the bi-directional manimpulation that can occurr. In other words, the user may interact with the media (as is right now) on both devices. A movie could be playing on the x86 renderer and the user can then scrub the track position back to the beginning, and it will update on the client device (e.g. phone/tablet). Pause and play will also reflect their states across both devices. Worth noting is that the AllCast application doesn't do this (the client and renderer applications can be downloaded). I am basically manimpulating the communication system to achieve this effect.

Another note on scalibility, the AllCast client applcation (phone) has the ability to connect to a media server. This means that if Plex or XBMC is running on the network, media can be easily served through the AllCast client app (or another). The idea is to be compatible with the OpenHome/UPnP/DLNA networks, but also support other popular applcations (AC has about 5 million downloads) Apple, Chrome, etc..